+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Support `drawing_area#paint(red, green, blue)` operation to set the initial `drawing_area` paint color base
 - Support `arc` as an operation inside `path` in Cairo declarative shape syntax
 - Support `arc_negative` cairo graphics shape
+- Support cairo graphics shape `clip`
 
 ## 0.0.4